Stop using super in action plugins
In a fit of epic inexplicability, the use of super for the run method in
action plugins can cause infinite recursion for code (such as Zuul) that
subclasses the action plugins. The issue shows up as the super call
finding the run method on itself, rather than on ActionBase, which of
course then results in the super running again and finding the run
method on itself ... and so on.

There is literally nothing understandable about WHY this occurs. But in
the case of the object hierarchy of the Action Plugins, calling
ActionBase.run is the correct method to call anyway, so there should be
no direct negative impact.